final commit added full referential integrity, reduced amount of classes, updated README, small other changes

+++ src/main/java/com/chabbay/dataobjects/AddressInformationController.java
@@ -2,10 +2,14 @@ package com.chabbay.dataobjects;
 
 import com.chabbay.dataobjects.objects.Address;
 import com.chabbay.dataobjects.objects.AddressInformation;
+import com.chabbay.dataobjects.objects.Country;
+import com.chabbay.dataobjects.objects.Customer;
 import com.chabbay.dataobjects.repositories.AddressInformationRepository;
 import com.chabbay.dataobjects.repositories.AddressRepository;
+import com.chabbay.dataobjects.repositories.CustomerRepository;
 import com.chabbay.errorhandling.exceptions.DataNotFoundException;
 import io.swagger.annotations.Api;
+import java.util.List;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -14,7 +18,6 @@ import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.IanaLinkRelations;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import java.util.List;
 
 /**
  * controller for the AddressInformation entity that defines endpoints
@@ -28,24 +31,37 @@ import java.util.List;
 @Api(tags="AddressInformation")
 public class AddressInformationController {
     private final AddressInformationRepository repository;
-    private final AddressInformationModelAssembler assembler;
     private final AddressRepository addressRepository;
+    private final CustomerRepository customerRepository;
+    private final AddressInformationModelAssembler assembler;
     private final String PATH = "/addressinformation";
 
-    public AddressInformationController(AddressInformationRepository repository, AddressInformationModelAssembler assembler, AddressRepository addressRepository) {
+    public AddressInformationController(AddressInformationRepository repository,
+                                        AddressInformationModelAssembler assembler, AddressRepository addressRepository,
+                                        CustomerRepository customerRepository) {
         this.repository = repository;
         this.assembler = assembler;
         this.addressRepository = addressRepository;
+        this.customerRepository = customerRepository;
     }
 
-    //select all
+    /**
+     * select all
+     *
+     * @Produces 200
+     */
     @GetMapping(PATH)
     CollectionModel<EntityModel<AddressInformation>> selectAll() {
         List<EntityModel<AddressInformation>> list = repository.findAll().stream().map(assembler::toModel).toList();
         return assembler.toCollection(list);
     }
 
-    //select
+    /**
+     * select
+     *
+     * @Produces 200
+     * @Produces 404 if not found
+     */
     @GetMapping(PATH + "/{id}")
     EntityModel<AddressInformation> select(@PathVariable Long id) {
         AddressInformation value = repository.findById(id).orElseThrow(() ->
@@ -53,14 +69,22 @@ public class AddressInformationController {
         return assembler.toModel(value);
     }
 
-    //insert
+    /**
+     * insert
+     *
+     * @Produces 201
+     */
     @PostMapping(PATH)
     ResponseEntity<?> insert(@RequestBody AddressInformation value) {
         EntityModel<AddressInformation> entityModel = assembler.toModel(repository.save(value));
         return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
     }
 
-    //update
+    /**
+     * update
+     *
+     * @Produces 201
+     */
     @PutMapping(PATH + "/{id}")
     ResponseEntity<?> update(@PathVariable Long id, @RequestBody AddressInformation value) {
         AddressInformation updated =  repository.findById(id).map(v -> {
@@ -77,17 +101,26 @@ public class AddressInformationController {
         return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
     }
 
-    //delete
+    /**
+     * delete
+     * block if referenced by Address or Customer
+     *
+     * @Produces 200
+     * @Produces 404 if not found
+     * @Produces 409 if referenced
+     */
     @DeleteMapping(PATH + "/{id}")
     ResponseEntity<?> delete(@PathVariable Long id) {
-        repository.deleteById(id);
-
-        //cascading delete of referenced Address entities
         List<Address> addresses = addressRepository.findByAddressInformationId(id);
-        for (Address address : addresses ) {
-            addressRepository.deleteById(address.getId());
+        List<Customer> customers = customerRepository.findByAddressInformationId(id);
+        if ((addresses.size() > 0) || (customers.size() > 0)) {
+            return ResponseEntity.status(409).build();
         }
 
+        repository.findById(id).orElseThrow(() ->
+                new DataNotFoundException(AddressInformation.class, id));
+
+        repository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
 }
